+++ plugins/postcss-focus-visible/.tape.mjs
@@ -11,6 +11,12 @@ postcssTape(plugin)({
 			replaceWith: '[data-focus-visible-added]'
 		}
 	},
+	'basic:disable-polyfill-ready-class': {
+		message: 'supports { disablePolyfillReadyClass: true } usage',
+		options: {
+			disablePolyfillReadyClass: true
+		}
+	},
 	'basic:preserve': {
 		message: 'supports { preserve: false } usage',
 		options: {
@@ -40,6 +46,12 @@ postcssTape(plugin)({
 		message: 'minimal example',
 		options: {
 			replaceWith: '[focus-visible]'
+		}
+	},
+	'examples/example:disable-polyfill-ready-class': {
+		message: 'minimal example',
+		options: {
+			disablePolyfillReadyClass: true
 		}
 	},
 });